refactor(tokenizers): extract model type detector into microcrate
