+++ crates/bitnet-tokenizer-model-core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "bitnet-tokenizer-model-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "Core model/vocabulary detection helpers for tokenizers"
+
+[dependencies]
+bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
+
