+++ crates/bitnet-tokenizers/Cargo.toml
@@ -20,6 +20,7 @@ include = [
 bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
 bitnet-models = { path = "../bitnet-models", version = "0.2.1-dev" }
 bitnet-quantization = { path = "../bitnet-quantization", version = "0.2.1-dev" }
+bitnet-tokenizer-model-core = { path = "../bitnet-tokenizer-model-core", version = "0.2.1-dev" }
 anyhow.workspace = true
 tracing.workspace = true
 serde = { workspace = true, features = ["derive"] }
